requirements.txt
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.7
cffi==1.12.3
colorama==0.4.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
Django==2.2.3
django-extensions==2.2.1
django-haystack==2.8.1
Fabric3==1.14.post1
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.15.1
jieba==0.39
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.17.0
paramiko==2.6.0
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.9
pycparser==2.19
Pygments==2.4.2
PyNaCl==1.3.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
wcwidth==0.1.7
Whoosh==2.7.4
